Software/Website/requirements.txt
flask_socketio
flask_wtf
opencv_python
pyserial
pigpio
WTForms
ctypesgen
simple_websocket
pylibmc
eventlet
flask_sqlalchemy
flask_migrate